third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-css-2-expected.txt

Verifies CSS pretty-printing functionality.


Running: testFontFace
====== 8< ------
@font-face {
    font-family: MyHelvetica;
    src: local('Helvetica Neue Bold'),local('HelveticaNeue-Bold'),url(MgOpenModernaBold.ttf);
    font-weight: bold;
}

div {
    color: red
}

------ >8 ======
Correct mapping for <font-face>
Correct mapping for <red>

Running: testCharsetRule
====== 8< ------
@charset 'iso-8859-15';p {
    margin: 0
}

------ >8 ======
Correct mapping for <charset>
Correct mapping for <iso>
Correct mapping for <margin>

Running: testImportRule
====== 8< ------
@import url('bluish.css') projection,tv;span {
    border: 1px solid black
}

------ >8 ======
Correct mapping for <import>
Correct mapping for <bluish>
Correct mapping for <projection>
Correct mapping for <span>
Correct mapping for <border>
Correct mapping for <black>

Running: testImportWithMediaQueryRule
====== 8< ------
@import url('landscape.css') screen and (orientation: landscape);
article {
    background: yellow
}

------ >8 ======
Correct mapping for <import>
Correct mapping for <url>
Correct mapping for <orientation>
Correct mapping for <article>
Correct mapping for <background>
Correct mapping for <yellow>

Running: testKeyframesRule
====== 8< ------
p {
    animation-duration: 3s;
}

@keyframes slidein {
    from {
        margin-left: 100%;
        width: 300%;
    }

    to {
        margin-left: 0%;
        width: 100%;
    }
}

p {
    animation-name: slidein
}

------ >8 ======
Correct mapping for <animation-duration>
Correct mapping for <3s>
Correct mapping for <keyframes>
Correct mapping for <from>
Correct mapping for <300%>
Correct mapping for <animation-name>
